Refactor gs_usb driver into modular components
